requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
MouseInfo==0.1.3
moviepy==1.0.3
mutagen==1.47.0
numpy==2.1.3
outcome==1.3.0.post0
pillow==11.0.0
proglog==0.1.10
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydub==0.25.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.32.3
rubicon-objc==0.4.9
selenium==4.26.1
setuptools==75.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
tinytag==2.0.0
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wheel==0.45.0
wsproto==1.2.0
